+++ changelog.md
@@ -1,5 +1,6 @@
 ### v4.7.8
 - Fix: lazy loading race condition in demos
+- Fix: image widening not always working on iOS when used with lazy loading
 
 ### v4.7.7
 - New: add "black" theme, which features a fully-black background to optimize for AMOLED displays.